Refactor: CurrenciesSelector: rename currencySelector to currenciesDataSelector
