+++ src/components/Currency/redux/currenciesDataSelector.ts
